Harden semantic language learning
