+++ tools/SemanticAtlasExperiment/manual-seeds.review.json
@@ -228,6 +228,116 @@
         "preserve food with Expanded Foods",
         "build a water wheel with Millwright"
       ]
+    },
+    {
+      "id": "everyday-conversation",
+      "name": "Everyday Conversation",
+      "reviewStatus": "approved-for-experiment",
+      "tags": ["manual", "roleplay", "everyday", "conversation"],
+      "sources": [],
+      "terms": [
+        "hello",
+        "goodbye",
+        "yes",
+        "no",
+        "maybe",
+        "please",
+        "thanks",
+        "sorry",
+        "name",
+        "friend",
+        "stranger",
+        "person",
+        "people",
+        "question",
+        "answer",
+        "truth",
+        "lie",
+        "story",
+        "rumor",
+        "promise",
+        "trust",
+        "peace",
+        "threat",
+        "fear",
+        "angry",
+        "happy",
+        "sad",
+        "tired"
+      ],
+      "phrases": [
+        "hello stranger what is your name",
+        "thank you for telling the truth",
+        "please answer my question",
+        "I am sorry for the mistake",
+        "we can speak in peace",
+        "do you trust these people",
+        "that story sounds like a rumor",
+        "I am tired and afraid tonight"
+      ]
+    },
+    {
+      "id": "everyday-survival-actions",
+      "name": "Everyday Survival Actions",
+      "reviewStatus": "approved-for-experiment",
+      "tags": ["manual", "roleplay", "everyday", "survival", "action"],
+      "sources": [],
+      "terms": [
+        "come",
+        "go",
+        "wait",
+        "stop",
+        "follow",
+        "run",
+        "hide",
+        "bring",
+        "carry",
+        "give",
+        "take",
+        "need",
+        "want",
+        "find",
+        "look",
+        "listen",
+        "eat",
+        "drink",
+        "sleep",
+        "rest",
+        "warm",
+        "cold",
+        "fire",
+        "water",
+        "food",
+        "medicine",
+        "wound",
+        "sick",
+        "lost",
+        "near",
+        "far",
+        "here",
+        "there",
+        "inside",
+        "outside",
+        "north",
+        "south",
+        "east",
+        "west",
+        "morning",
+        "night",
+        "today",
+        "tomorrow",
+        "later"
+      ],
+      "phrases": [
+        "come here and follow me",
+        "wait outside until morning",
+        "run north and hide near the fire",
+        "bring food and water for the sick traveler",
+        "I need medicine for this wound",
+        "take the supplies and carry them inside",
+        "we are lost far from camp",
+        "sleep now and leave tomorrow"
+      ]
     }
   ]
 }
